[hotfix] [tableAPI] Moved tests to correct package.
