+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/api/table/typeutils/RowComparatorTest.scala
@@ -16,16 +16,15 @@
  * limitations under the License.
  */
 
-package org.apache.flink.api.scala.typeutils
+package org.apache.flink.api.table.typeutils
 
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo
 import org.apache.flink.api.common.typeutils.{ComparatorTestBase, TypeComparator, TypeSerializer}
 import org.apache.flink.api.java.tuple
 import org.apache.flink.api.java.typeutils.{TupleTypeInfo, TypeExtractor}
-import org.apache.flink.api.scala.typeutils.RowComparatorTest.MyPojo
 import org.apache.flink.api.table.Row
-import org.apache.flink.api.table.typeutils.RowTypeInfo
+import org.apache.flink.api.table.typeutils.RowComparatorTest.MyPojo
 import org.junit.Assert._
 
 class RowComparatorTest extends ComparatorTestBase[Row] {
@@ -134,3 +133,4 @@ object RowComparatorTest {
     }
   }
 }
+
